добавлены комментарии по действиям при решении задач

+++ src/Main.java
@@ -25,7 +25,7 @@ public static  void task1() {
         System.out.println("Задача 2");
     float a=27.12f;
     System.out.println(a);
-    long b = 987678965549l;
+    long b = 987678965549L;
     System.out.println(b);
     float c = 2.786f;
     System.out.println(c);
@@ -39,41 +39,41 @@ public static  void task1() {
     System.out.println(h);}
     public static  void task3() {
         System.out.println("Задача 3");
-    byte LudPavl = 23;
-    byte AnSerg = 27;
-    byte EkatAndr = 30;
-    short copybook = 480;
-    byte totalStudent = (byte) (LudPavl + AnSerg + EkatAndr);
-    byte copybookOnStudent = (byte) (copybook/totalStudent);
+    byte LudPavl = 23; // инициализация количества учеников в классе
+    byte AnSerg = 27; // инициализация количества учеников в классе
+    byte EkatAndr = 30; // инициализация количества учеников в классе
+    short copybook = 480; // инициализация количество тетрадей
+    byte totalStudent = (byte) (LudPavl + AnSerg + EkatAndr); // расчет количества всех учеников
+    byte copybookOnStudent = (byte) (copybook/totalStudent); // расчет количества тетрадей на 1 ученика
         System.out.println("На каждого ученика рассчитано " + copybookOnStudent + " листов бумаги.");}
     public static  void task4() {
         System.out.println("Задача 4");
-    byte bottle = 16;
-    byte timeMin = 2;
+    byte bottle = 16; // инициализация количество бутылок произведенных за 2 минуты
+    byte timeMin = 2; // инициализация количества времени для производства 16 бутылок
     byte performanceBottleTime = (byte) (bottle/timeMin); //расчет производительности за 1 минуту
-    byte OneSpecifiedIntervalMinute = 20;
-    byte TwoSpecifiedIntervalDay = 1;
-    byte ThreeSpecifiedIntervalDay = 3;
-    byte FourSpecifiedIntervalMonth = 1;
+    byte OneSpecifiedIntervalMinute = 20; // инициализация количества времени в 1 временном интервале
+    byte TwoSpecifiedIntervalDay = 1; // инициализация количества времени во 2 временном интервале
+    byte ThreeSpecifiedIntervalDay = 3; // инициализация количества времени в 3 временном интервале
+    byte FourSpecifiedIntervalMonth = 1; // инициализация количества времени в 4 временном интервале
     byte minuteHour = 60; // количество минут в часе
     byte hourDay = 24; // количество часов в сутках
     byte dayMonth = 30; // количество дней в месяце
-    short OnePerformanceBottle = (short) (performanceBottleTime * OneSpecifiedIntervalMinute);
+    short OnePerformanceBottle = (short) (performanceBottleTime * OneSpecifiedIntervalMinute); // расчет количества произведенных бутылок
     System.out.println("За " + OneSpecifiedIntervalMinute + " минут машина произвела " + OnePerformanceBottle + " штук бутылок.");
-    short TwoSpecifiedIntervalMinute = (short) (TwoSpecifiedIntervalDay * hourDay * minuteHour);
-    short TwoPerformanceBottle = (short) (TwoSpecifiedIntervalMinute * performanceBottleTime);
+    short TwoSpecifiedIntervalMinute = (short) (TwoSpecifiedIntervalDay * hourDay * minuteHour); // расчет количества минут в заданном интервале времени
+    short TwoPerformanceBottle = (short) (TwoSpecifiedIntervalMinute * performanceBottleTime); // расчет количества произведенных бутылок
     System.out.println("За " + TwoSpecifiedIntervalDay + " сутки машина произвела " + TwoPerformanceBottle + " штук бутылок.");
-    short ThreeSpecifiedIntervalMinute = (short) (ThreeSpecifiedIntervalDay * hourDay * minuteHour);
-    int ThreePerformanceBottle = (short) (ThreeSpecifiedIntervalMinute * performanceBottleTime);
+    short ThreeSpecifiedIntervalMinute = (short) (ThreeSpecifiedIntervalDay * hourDay * minuteHour); // расчет количества минут в заданном интервале времени
+    int ThreePerformanceBottle = (short) (ThreeSpecifiedIntervalMinute * performanceBottleTime); // расчет количества произведенных бутылок
     System.out.println("За " + ThreeSpecifiedIntervalDay + " дня машина произвела " + ThreePerformanceBottle + " штук бутылок.");
-    int FourSpecifiedIntervalMinute = (short) (FourSpecifiedIntervalMonth * dayMonth * hourDay * minuteHour);
-    int FourPerformanceBottle = (short) (FourSpecifiedIntervalMinute * performanceBottleTime);
+    int FourSpecifiedIntervalMinute = (short) (FourSpecifiedIntervalMonth * dayMonth * hourDay * minuteHour); // расчет количества минут в заданном интервале времени
+    int FourPerformanceBottle = (short) (FourSpecifiedIntervalMinute * performanceBottleTime); // расчет количества произведенных бутылок
     System.out.println("За " + FourSpecifiedIntervalMonth + " месяц машина произвела " + FourPerformanceBottle + " штук бутылок.");}
     public static  void task5() {
         System.out.println("Задача 5");
-    byte potTotal = 120;
-    byte oneClassPotWhite = 2;
-    byte oneClassPotBrown = 4;
+    byte potTotal = 120; // инициализация потребности количества банок красок
+    byte oneClassPotWhite = 2; // инициализация количества банок белой краски на 1 класс
+    byte oneClassPotBrown = 4; // инициализация количества банок коричневой краски на 1 класс
     byte oneClassPotTotal = (byte) (oneClassPotBrown + oneClassPotWhite); // Сколько банок красок уходит на 1 класс
     byte totalClassSchool = (byte) (potTotal / oneClassPotTotal); // Всего классов в школе
     byte potTotalWhite = (byte) (totalClassSchool * oneClassPotWhite); // сколько банок белой краски на 1 класс
@@ -81,20 +81,20 @@ public static  void task1() {
     System.out.println("В школе, где " + totalClassSchool + " классов, нужно " + potTotalWhite + " банок белой краски и " + potTotalBrown + " банок коричневой краски.");}
     public static  void task6() {
         System.out.println("Задача 6");
-    byte quantityBanana = 5;
-    byte weightBananaGrams = 80;
-    float weightBananaGramsTotal = (float) (quantityBanana * weightBananaGrams);
-    short volumeMilkMl = 200;
-    float gramsInMl = 1.05f;
-    float weightMilkGramsTotal = (float) (volumeMilkMl * gramsInMl);
-    byte quantityIceCreamSundae = 2;
-    byte weightIceCreamSundaeGrams = 100;
-    float weightIceCreamSundaeTotal = (float) (quantityIceCreamSundae * weightIceCreamSundaeGrams);
-    byte quantityEgg = 4;
-    byte weightEggGrams = 70;
-    float weightEggGramsTotal = (float) (quantityEgg * weightEggGrams);
-    float weightBreakfastTotalGrams = (float) (weightBananaGramsTotal + weightMilkGramsTotal + weightIceCreamSundaeTotal+ weightEggGramsTotal);
-    short gramsInKilogram = 1000;
-    float weightBreakfastTotalKg = (float) (weightBreakfastTotalGrams / gramsInKilogram);
+    byte quantityBanana = 5; // инициализация количества бананов
+    byte weightBananaGrams = 80; // инициализация веса одного банана
+    float weightBananaGramsTotal = (float) (quantityBanana * weightBananaGrams); // расчет веса всех бананов
+    short volumeMilkMl = 200; // инициализация количества молока в мл
+    float gramsInMl = 1.05f; // инициализация количество грамм в 1 мл
+    float weightMilkGramsTotal = volumeMilkMl * gramsInMl; // расчет веса молока
+    byte quantityIceCreamSundae = 2; // инициализация количество брикетов мороженного
+    byte weightIceCreamSundaeGrams = 100; // инициализация веса 1 брикета мороженного
+    float weightIceCreamSundaeTotal = (float) (quantityIceCreamSundae * weightIceCreamSundaeGrams); // расчет веса всего мороженного
+    byte quantityEgg = 4; // инициализация количества яиц
+    byte weightEggGrams = 70; // инициализация веса 1 яйца
+    float weightEggGramsTotal = (float) (quantityEgg * weightEggGrams); // расчет веса всего мороженного
+    float weightBreakfastTotalGrams = (weightBananaGramsTotal + weightMilkGramsTotal + weightIceCreamSundaeTotal+ weightEggGramsTotal); // расчет веса всего завтрака
+    short gramsInKilogram = 1000; // количество грамм в 1 кг
+    float weightBreakfastTotalKg =  (weightBreakfastTotalGrams / gramsInKilogram); // перевод веса из грамм в кг
     System.out.println(" Вес завтрака спортсмена составляет " + weightBreakfastTotalGrams + " грамма или " + weightBreakfastTotalKg + " килограмм.");}
 }
